made sure when we delete image it also sets the poster/picture to null for profile

+++ app/src/main/java/ca/cmput301t05/placeholder/database/tables/ImageDetailTable.java
@@ -3,7 +3,9 @@ package ca.cmput301t05.placeholder.database.tables;
 import android.util.Log;
 
 import com.google.firebase.firestore.CollectionReference;
+import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FieldValue;
 import com.google.firebase.storage.StorageReference;
 
 import ca.cmput301t05.placeholder.database.DatabaseManager;
@@ -42,31 +44,62 @@ public class ImageDetailTable extends Table<ImageDetails>{
 
         StorageReference storageRef = DatabaseManager.getInstance().getStorage().getReference().child(imageDetails.getImagePath());
 
-        storageRef.delete().addOnSuccessListener(unused -> {
+        //grab the object key so we can reset it
 
-            //now delete imagedetails
-            this.deleteDocument(imageDetails.getId(), new DocumentCallback() {
-                @Override
-                public void onSuccess(Object document) {
-                    Log.d("Image Database", "Deleted Image");
-                    callback.onSuccess(document);
+
+        String [] objectID = imageDetails.getObjectID().split(":");
+        DocumentReference docRef;
+        String path;
+
+        if (objectID[0].equals("Profile")) {
+            docRef = DatabaseManager.getInstance().getDb().collection("profiles").document(objectID[1]);
+            path = "profilePictureID";
+        } else {
+            docRef = DatabaseManager.getInstance().getDb().collection("events").document(objectID[1]);
+            path = "eventPosterID";
+        }
+
+        docRef.update(path, FieldValue.delete()).addOnCompleteListener(task -> {
+
+            if (task.isSuccessful()){
+
+                //deleting the picture in firebase storage
+                storageRef.delete().addOnSuccessListener(unused -> {
+
+                    //now delete imagedetails
+                    this.deleteDocument(imageDetails.getId(), new DocumentCallback() {
+                        @Override
+                        public void onSuccess(Object document) {
+                            Log.d("Image Database", "Deleted Image");
+                                callback.onSuccess(document);
+                        }
+
+                        @Override
+                        public void onFailure(Exception e) {
+                            callback.onFailure(e);
+                            }
+                        });
+
+
+                    }).addOnFailureListener(e -> {
+                        // If the image ID is invalid or the image does not exist
+                        Log.d("Image Database", "Error: " + e.getMessage());
+                    });
+
                 }
 
-                @Override
-                public void onFailure(Exception e) {
-                    callback.onFailure(e);
-
+                else {
+                    callback.onFailure(task.getException());
                 }
+
+
             });
+        }
 
 
-        }).addOnFailureListener(e -> {
-            // If the image ID is invalid or the image does not exist
-            Log.d("Image Database", "Error: " + e.getMessage());
-        });
 
 
     }
 
 
-}
+
